+++ tools/container_environment/startup.sh
@@ -1,0 +1,3 @@
+cd /app
+npm install
+node --eval "setInterval(() => {}, 10000);"